Remove Jsch and add Apache Mina

+++ datasophon-api/src/main/java/com/datasophon/api/utils/MinaUtils.java
@@ -1,28 +1,19 @@
 package com.datasophon.api.utils;
 
-import ch.ethz.ssh2.SFTPv3Client;
-import ch.ethz.ssh2.Session;
-import com.jcraft.jsch.ChannelSftp;
-import com.jcraft.jsch.JSchException;
-import com.jcraft.jsch.SftpATTRS;
-import com.jcraft.jsch.SftpException;
 import org.apache.sshd.client.SshClient;
 import org.apache.sshd.client.channel.ChannelExec;
 import org.apache.sshd.client.channel.ClientChannelEvent;
 import org.apache.sshd.client.session.ClientSession;
-import org.apache.sshd.client.subsystem.SubsystemClient;
-import org.apache.sshd.sftp.client.SftpClient;
 import org.apache.sshd.sftp.client.SftpClientFactory;
 import org.apache.sshd.sftp.client.fs.SftpFileSystem;
 import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.NoSuchAlgorithmException;
 import java.util.EnumSet;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
@@ -30,60 +21,85 @@ import java.util.concurrent.TimeUnit;
 public class MinaUtils {
     private static final org.slf4j.Logger LOG = LoggerFactory.getLogger(MinaUtils.class);
 
+    private String sshHost;
+    private Integer sshPort;
+    private String sshUser;
+    private String privateKey;
+
+    private SshClient sshClient;
+    private ClientSession session;
+    private SftpFileSystem sftp;
+
+    public MinaUtils(String sshHost, Integer sshPort, String sshUser, String privateKey) {
+        this.sshHost = sshHost;
+        this.sshPort = sshPort;
+        this.sshUser = sshUser;
+        this.privateKey = privateKey;
+    }
+
 
     /**
      * 打开远程会话
-     *
-     * @param sshHost 主机
-     * @param sshPort 端口
-     * @param sshUser 用户名，如果为null，默认root
-     * @param sshPass 密码
-     * @return {@link Session}
      */
-    public static ClientSession openConnection(String sshHost, Integer sshPort, String sshUser, String sshPass) {
+    public ClientSession openConnection() {
 
-        SshClient client = SshClient.setUpDefaultClient();
-        client.start();
-        ClientSession session = null;
+        this.sshClient = SshClient.setUpDefaultClient();
+        this.sshClient.start();
         try {
-            session = client.connect(sshUser, sshHost, sshPort).verify().getClientSession();
-
-            session.addPasswordIdentity(sshPass);
-            if (session.auth().verify(3000).isFailure()) {
+            this.session = this.sshClient.connect(sshUser, sshHost, sshPort).verify().getClientSession();
+            this.session.addPublicKeyIdentity(getKeyPairFromString(privateKey));
+            if (this.session.auth().verify().isFailure()) {
                 LOG.info("验证失败");
                 return null;
             }
+            this.sftp = SftpClientFactory.instance().createSftpFileSystem(this.session);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
         LOG.info(sshHost + " 连接成功");
-        return session;
+        return this.session;
     }
 
     /**
-     * 开启exec通道
-     *
-     * @param session Session
-     * @return ChanelExec
+     * 关闭远程会话
      */
-    public static ChannelExec openChannelExec(ClientSession session) {
-        ChannelExec channelExec = null;
+    public void closeConnection() {
         try {
-            channelExec = session.createExecChannel("exec");
+            this.session.close();
+            this.sftp.close();
+            this.sshClient.stop();
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-        return channelExec;
     }
+
+    /**
+     * 获取密钥对
+     */
+    static KeyPair getKeyPairFromString(String pk) {
+        final KeyPairGenerator rsa;
+        try {
+            rsa = KeyPairGenerator.getInstance("RSA");
+            final KeyPair keyPair = rsa.generateKeyPair();
+            final ByteArrayOutputStream stream = new ByteArrayOutputStream();
+            stream.write(pk.getBytes());
+            final ObjectOutputStream o = new ObjectOutputStream(stream);
+            o.writeObject(keyPair);
+            return keyPair;
+        } catch (NoSuchAlgorithmException | IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
 
     /**
      * 同步执行,需要获取执行完的结果
      *
-     * @param session Session
      * @param command 命令
      * @return 结果
      */
-    public static String execCmdWithResult(ClientSession session, String command, long timeout) {
+    public String execCmdWithResult(String command) {
+        this.openConnection();
         LOG.info("exe cmd: {}", command);
         // 命令返回的结果
         ChannelExec ce = null;
@@ -92,19 +108,18 @@ public class MinaUtils {
         // 错误信息
         ByteArrayOutputStream err = new ByteArrayOutputStream();
         try {
-            ce = session.createExecChannel(command);
+            ce = this.session.createExecChannel(command);
             ce.setOut(out);
             ce.setErr(err);
-            // Execute and wait
+            // 执行并等待
             ce.open();
             Set<ClientChannelEvent> events =
-                    ce.waitFor(EnumSet.of(ClientChannelEvent.CLOSED), TimeUnit.SECONDS.toMillis(timeout));
-            //  Check if timed out
+                    ce.waitFor(EnumSet.of(ClientChannelEvent.CLOSED), TimeUnit.SECONDS.toMillis(10000));
+            //  检查请求是否超时
             if (events.contains(ClientChannelEvent.TIMEOUT)) {
                 throw new Exception("mina 连接超时");
             }
-            session.close(false);
-
+            this.session.close(false);
             int exitStatus = ce.getExitStatus();
             LOG.info("mina result {}", exitStatus);
             if (exitStatus == 1) {
@@ -123,49 +138,40 @@ public class MinaUtils {
             }
         }
         LOG.info("exe cmd return : {}", out);
-        return out.toString();
-    }
-
-
-    /**
-     * 开启SFTP通道
-     *
-     * @param session Session
-     * @return ChannelSftp
-     * @throws Exception
-     */
-    public static SftpFileSystem openChannelSftp(ClientSession session) {
-        SftpFileSystem sftpFileSystem = null;
-        try {
-            sftpFileSystem = SftpClientFactory.instance().createSftpFileSystem(session);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        return sftpFileSystem;
+        return out.toString().trim();
     }
 
     /**
      * 上传文件,相同路径ui覆盖
      *
-     * @param sftp       sftp
      * @param remotePath 远程目录地址
      * @param inputFile  文件 File
      */
-    public static boolean uploadFile(SftpFileSystem sftp, String remotePath, String inputFile) {
+    public boolean uploadFile(String remotePath, String inputFile) {
+        this.openConnection();
         File uploadFile = new File(inputFile);
-        FileInputStream input = null;
-        Path path = sftp.getDefaultDir().resolve(remotePath);
+        InputStream input = null;
         try {
-            input = new FileInputStream(uploadFile);
+            Path path = this.sftp.getDefaultDir().resolve(remotePath);
             if (!Files.exists(path)) {
+                LOG.info("create pathHome {} ", path);
                 Files.createDirectories(path);
             }
+            input = Files.newInputStream(uploadFile.toPath());
             Path file = path.resolve(uploadFile.getName());
-            Files.deleteIfExists(file);
+            if (Files.exists(file)){
+                Files.deleteIfExists(file);
+            }
             Files.copy(input, file);
             return true;
         } catch (IOException e) {
             throw new RuntimeException(e);
+        } finally {
+            try {
+                this.sftp.close();
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         }
     }
 
@@ -173,12 +179,12 @@ public class MinaUtils {
      * 创建目录
      *
      * @param path
-     * @param sftp
      * @return
      */
-    public static boolean createDir(String path, SftpFileSystem sftp) {
+    public boolean createDir(String path) {
+        this.openConnection();
         try {
-            Path remoteRoot = sftp.getDefaultDir().resolve(path);
+            Path remoteRoot = this.sftp.getDefaultDir().resolve(path);
             if (!Files.exists(remoteRoot)) {
                 Files.createDirectories(remoteRoot);
                 return true;
@@ -190,17 +196,16 @@ public class MinaUtils {
     }
 
 
-    public static void main(String[] args) throws IOException {
-        ClientSession session = MinaUtils.openConnection("localhost", 22, "liuxin", "960319");
-        SftpFileSystem sftpFileSystem = openChannelSftp(session);
-//        String ls = execCmdWithResult(session, "java -version 2>&1 | sed '1!d' | sed -e 's/\"//g' | awk '{print $3}'", 1000);
-//        System.out.println(ls);
-//        session.close();
-
-//        boolean dir = MinaUtils.createDir("/Users/liuxin/opt/test", sftpFileSystem);
+    public static void main(String[] args) throws IOException, InterruptedException {
+        MinaUtils minaUtils = new MinaUtils("localhost", 22, "liuxin",
+                "/Users/liuxin/.ssh/id_rsa");
+            String ls = minaUtils.execCmdWithResult("arch");
+            System.out.println(ls);
+//        minaUtils.closeConnection();
+//        boolean dir = minaUtils.createDir("/home/shinow/test/");
 //        System.out.println(dir);
-        boolean uploadFile = uploadFile(sftpFileSystem, "/Users/liuxin/opt/test", "/Users/liuxin/opt/datax.tar.gz");
-        System.out.println(uploadFile);
+//        boolean uploadFile = minaUtils.uploadFile("/Users/liuxin/opt/test", "/Users/liuxin/Downloads/yarn-default.xml");
+//        System.out.println(uploadFile);
     }
 
 }
